+++ packages/app-project/src/components/AuthModal/shared/components/FieldLabel/FieldLabel.js
@@ -1,9 +1,0 @@
-import { SpacedText } from '@zooniverse/react-components'
-
-function FieldLabel (props) {
-  return (
-    <SpacedText weight='bold' {...props} />
-  )
-}
-
-export default FieldLabel