+++ packages/lib-react-components/src/ZooHeader/components/MainNavList/MainNavList.js
@@ -3,9 +3,9 @@ import NavListItem from '../NavListItem'
 import { Box } from 'grommet'
 
 export default function MainNavList({
+  adminMode = false,
   adminNavLinkLabel,
   adminNavLinkURL,
-  isAdmin = false,
   isNarrow = false,
   mainHeaderNavListLabels,
   mainHeaderNavListURLs
@@ -21,7 +21,7 @@ export default function MainNavList({
           url={url}
         />
       ))}
-      {isAdmin &&
+      {adminMode &&
         <NavListItem
           label={adminNavLinkLabel}
           url={adminNavLinkURL}
@@ -31,9 +31,9 @@ export default function MainNavList({
 }
 
 MainNavList.propTypes = {
+  adminMode: PropTypes.bool,
   adminNavLinkLabel: PropTypes.string.isRequired,
   adminNavLinkURL: PropTypes.string.isRequired,
-  isAdmin: PropTypes.bool,
   isNarrow: PropTypes.bool,
   mainHeaderNavListLabels: PropTypes.arrayOf(PropTypes.string).isRequired,
   mainHeaderNavListURLs: PropTypes.arrayOf(PropTypes.string).isRequired,